+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog 
 
 ## unreleased
-
+- CO-2130 - support Unzer Plugin
 
 ## 2.16.0 _2022-09-13_
 - CO-2113 - add TaxGroups to GlobalData Pull